# Use Python built-in str.lower in preference to asciiUpper2Lower character table translation