--- a/html5lib/html5parser.py
+++ b/html5lib/html5parser.py
@@ -11,7 +11,7 @@ from .treebuilders.base import Marker
 
 from . import _utils
 from .constants import (
-    spaceCharacters, asciiUpper2Lower,
+    spaceCharacters,
     specialElements, headingElements, cdataElements, rcdataElements,
     tokenTypes, tagTokenTypes,
     namespaces,
@@ -183,8 +183,7 @@ class HTMLParser(object):
         if (element.name == "annotation-xml" and
                 element.namespace == namespaces["mathml"]):
             return ("encoding" in element.attributes and
-                    element.attributes["encoding"].translate(
-                        asciiUpper2Lower) in
+                    element.attributes["encoding"].lower() in
                     ("text/html", "application/xhtml+xml"))
         else:
             return (element.namespace, element.name) in htmlIntegrationPointElements
@@ -520,7 +519,7 @@ def getPhases(debug):
             self.tree.insertDoctype(token)
 
             if publicId != "":
-                publicId = publicId.translate(asciiUpper2Lower)
+                publicId = publicId.lower()
 
             if (not correct or token["name"] != "html" or
                     publicId.startswith(
@@ -1165,7 +1164,7 @@ def getPhases(debug):
             framesetOK = self.parser.framesetOK
             self.startTagVoidFormatting(token)
             if ("type" in token["data"] and
-                    token["data"]["type"].translate(asciiUpper2Lower) == "hidden"):
+                    token["data"]["type"].lower() == "hidden"):
                 # input type=hidden doesn't change framesetOK
                 self.parser.framesetOK = framesetOK
 
@@ -1771,7 +1770,7 @@ def getPhases(debug):
 
         def startTagInput(self, token):
             if ("type" in token["data"] and
-                    token["data"]["type"].translate(asciiUpper2Lower) == "hidden"):
+                    token["data"]["type"].lower() == "hidden"):
                 self.parser.parseError("unexpected-hidden-input-in-table")
                 self.tree.insertElement(token)
                 # XXX associate with form
@@ -2512,11 +2511,11 @@ def getPhases(debug):
         def processEndTag(self, token):
             nodeIndex = len(self.tree.openElements) - 1
             node = self.tree.openElements[-1]
-            if node.name.translate(asciiUpper2Lower) != token["name"]:
+            if node.name.lower() != token["name"]:
                 self.parser.parseError("unexpected-end-tag", {"name": token["name"]})
 
             while True:
-                if node.name.translate(asciiUpper2Lower) == token["name"]:
+                if node.name.lower() == token["name"]:
                     # XXX this isn't in the spec but it seems necessary
                     if self.parser.phase == self.parser.phases["inTableText"]:
                         self.parser.phase.flushCharacters()